+++ app/(protected)/dashboard/page.js
@@ -43,10 +43,8 @@ export default async function DashboardPage() {
 
   return (
     <Dashboard
-      likedNames={[]}
       namesCreated={namesCreated}
       createdDescriptions={createdDescriptions}
-      likedDescriptions={[]}
     />
   );
 }